BlueCoveBlueZ_L2CAP: clear localAddr before use BlueCoveBlueZ_L2CAPServer: clear localAddr before use
From my message to bug reporters on the bluecove google code site:
The other attachment (init.diff) is a patch (which may already be
applied in the latest cvs), to init localAddr structs. Without this,
running on java8 will cause the l2cap server to fail to obtain a PSM
(port). Please also consider making this available.